stop forcing ordering of tokens df columns
In many cases it may be useful to include additional ids
in the leading columns of a tokens data frame, such as
paragraph and sentence identifiers. It makes sense that
these would follow the doc_id rather than coming after the
tokens column. This commit allows for this more flexible
format